WIP: rename app `game` to `pong_game`
