+++ backend/pong_game/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class GameConfig(AppConfig):
 	default_auto_field = 'django.db.models.BigAutoField'
-	name = 'game'
+	name = 'pong_game'
