+++ backend/pong_game/scripts/game/cli_update_score.py
