+++ backend/pong_game/templates/game/game_details.html
