+++ backend/transcendence/asgi.py
@@ -8,11 +8,12 @@ https://docs.djangoproject.com/en/5.1/howto/deployment/asgi/
 """
 
 import os
-from django.core.asgi import get_asgi_application
+
 from channels.auth import AuthMiddlewareStack
 from channels.routing import ProtocolTypeRouter, URLRouter
 from chat import routing as chat_routing
-from game import routing as game_routing
+from django.core.asgi import get_asgi_application
+from pong_game import routing as pong_routing
 from quiz import routing as quiz_routing
 from users import routing as user_status_routing
 
@@ -24,7 +25,7 @@ application = ProtocolTypeRouter(
 		'websocket': AuthMiddlewareStack(
 			URLRouter(
 				chat_routing.websocket_urlpatterns
-				+ game_routing.websocket_urlpatterns
+				+ pong_routing.websocket_urlpatterns
 				+ quiz_routing.websocket_urlpatterns
 				+ user_status_routing.websocket_urlpatterns,
 			)
